Add types through JSDoc
