+++ src/errors/ArgumentError.js
@@ -1,4 +1,5 @@
 export class ArgumentError extends Error {
+  /** @param {string} message */
   constructor(message) {
     super(message);
     this.name = 'ArgumentError';
