+++ src/errors/JwksError.js
@@ -1,4 +1,5 @@
 export class JwksError extends Error {
+  /** @param {string} message */
   constructor(message) {
     super(message);
     this.name = 'JwksError';
