+++ src/errors/JwksRateLimitError.js
@@ -1,4 +1,5 @@
 export class JwksRateLimitError extends Error {
+  /** @param {string} message */
   constructor(message) {
     super(message);
     this.name = 'JwksRateLimitError';
