+++ src/errors/SigningKeyNotFoundError.js
@@ -1,4 +1,5 @@
 export class SigningKeyNotFoundError extends Error {
+  /** @param {string} message */
   constructor(message) {
     super(message);
     this.name = 'SigningKeyNotFoundError';
