+++ src/integrations/express.js
@@ -1,7 +1,12 @@
+/** @typedef {import('../types.js').ExpressJwtOptions} ExpressJwtOptions */
+/** @typedef {import('../types.js').GetVerificationKey} GetVerificationKey */
+/** @typedef {import('../types.js').SecretCallbackLong} SecretCallbackLong */
+
 import { ArgumentError } from '../errors/ArgumentError.js';
 import { JwksClient } from '../JwksClient.js';
 import { allowedSignatureAlg } from './config.js';
 
+/** @type {ExpressJwtOptions['handleSigningKeyError']} */
 const handleSigningKeyError = (err, cb) => {
   // If we didn't find a match, can't provide a key.
   if (err && err.name === 'SigningKeyNotFoundError') {
@@ -14,6 +19,10 @@ const handleSigningKeyError = (err, cb) => {
   }
 };
 
+/**
+ * @param {ExpressJwtOptions} options
+ * @returns {SecretCallbackLong | GetVerificationKey}
+ */
 export function expressJwtSecret(options) {
   if (options === null || options === undefined) {
     throw new ArgumentError('An options object must be provided when initializing expressJwtSecret');
