+++ src/wrappers/callbackSupport.js
@@ -1,5 +1,8 @@
+/** @typedef {import('../JwksClient.js').JwksClient} JwksClient */
+
 import { callbackify } from 'node:util';
 
+/** @param {JwksClient} client */
 export const callbackSupport = (client) => {
   const getSigningKey = client.getSigningKey.bind(client);
