+++ src/wrappers/interceptor.js
@@ -1,8 +1,14 @@
+/** @typedef {import('../JwksClient.js').JwksClient} JwksClient */
+/** @typedef {import('../types.js').Options} Options */
+
 import { retrieveSigningKeys } from '../utils.js';
 
 /**
- * Uses getKeysInterceptor to allow users to retrieve keys from a file,
- * external cache, or provided object before falling back to the jwksUri endpoint
+ * Uses getKeysInterceptor to allow users to retrieve keys from a file, external cache, or provided object before
+ * falling back to the jwksUri endpoint
+ *
+ * @param {JwksClient} client
+ * @param {Omit<Options, 'getKeysInterceptor'> & Required<Pick<Options, 'getKeysInterceptor'>>} options
  */
 export function getKeysInterceptor(client, { getKeysInterceptor }) {
   const getSigningKey = client.getSigningKey.bind(client);
